Fixed moonfire links being lunar inspiration
Technically not 'wrong' as that's how it goes in the original compendium as well, but I think it's unintuitive for readers. 

The text shows moonfire(and that's the button they should press), but when you hover it shows tooltip for a talent that enables moonfire in cat form, not the actual spell. It's safe from unshift issue since there's a checkbox for lunar inspiration talented, so the readers without LI would not accidentally get the idea of pressing moonfire.