Balance APL: use built-in expressions instead of variables
Instead of defining a new variable, use the already-defined (in SimulationCraft) `eclipse.X` expressions. This change should have no effect on the rotation. One could also be built in SimC to replace `variable.eclipse_remains`, but I will leave this PR at the existing expressions.

While I understand if you don't wish to make this change, it is a big help to the Hekili project which is downstream of this APL-wise.